requirements.txt
APScheduler==3.6.3
asgiref==3.11.1
cachetools==4.2.2
certifi==2026.2.25
Django==4.2.12
django-environ==0.13.0
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-telegram-bot==13.15
pytz==2026.1.post1
redis==7.3.0
six==1.17.0
sqlparse==0.5.5
tornado==6.1
tzlocal==5.3.1
